pyproject.toml
[tool.poetry]
name = "everytime-cleaner"
version = "0.1.0"
description = ""
authors = ["SonJinHYo <wlsgy95@naver.com>"]
license = "\"MIT\""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bs4 = "^0.0.2"
selenium = "^4.21.0"
webdriver-manager = "^4.0.1"
openai = "^1.30.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
